- New clonezilla image with support for network install.

+++ image/clonezilla-overwrites/bibos/install-local.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
-mkdir /home/partimag/bibos
-sudo mount --bind /live/image/bibos-image /home/partimag/bibos
+mkdir /home/partimag/
+sudo mount --bind /live/image/bibos-images /home/partimag/
 
 DEV=`sudo /usr/bin/perl /live/image/bibos/find_disk.pl`
 if [ "" == "$DEV" ]; then
@@ -22,6 +22,12 @@ if [ $? -ne 0 ]; then
     exit 1;
 fi
 
-sudo /usr/bin/perl /live/image/bibos/clone_image.pl $DEV
+IMAGE=`sudo /usr/bin/perl /live/image/bibos/select_image.pl`;
+if [ $? -ne 0 ]; then
+    echo "No image selected, exiting";
+    exit 1;
+fi
+
+sudo /usr/bin/perl /live/image/bibos/clone_image.pl $DEV $IMAGE
 
 sudo reboot
